Auth\BasicTest: split constructor specific tests to dedicated test class
The other tests in the `BasicTest` class are integration tests, not unit tests, so should remain in the generic class test class.
